requirements.txt
# Existing packages
streamlit==1.29.0
pandas==2.1.4
numpy==1.24.3
hdbcli==2.19.21
scikit-learn==1.3.2
google-generativeai==0.3.2
langgraph==0.0.45

# MCP packages
fastmcp==0.2.0
httpx==0.27.0
python-dotenv==1.0.0